Ajoout du système d'arbre de construction pour la scène ainsi que le vieillissement

+++ AppTinyMesh/Include/surface.hpp
@@ -20,7 +20,7 @@ class Surface {
     inline const Vector& pointC(const size_t x, const size_t y) const;
 
 
-    int height, width, numberControlPointH, numberControlPointW;
+    size_t height, width, numberControlPointH, numberControlPointW;
 
 };
