+++ AppTinyMesh/Source/bezier.cpp
@@ -37,8 +37,8 @@ Vector Bezier::calcul_normal(const float u, const float v) const {
 Vector Bezier::calcul_vertex(const float u, const float v) const {
     Vector res = Vector(0);
 
-    for (int i = 0; i < numberControlPointW; i++) {
-        for(int j = 0; j < numberControlPointH; j++) {
+    for (size_t i = 0; i < numberControlPointW; i++) {
+        for(size_t j = 0; j < numberControlPointH; j++) {
             res += bernstein(numberControlPointW, i, u)
                 * bernstein(numberControlPointH, j, v)
                 * pointC(i, j);
